refactor: extract PageSelection component to new file
